third_party/WebKit/LayoutTests/media/track/track-css-user-settings-override-internal-settings-expected.txt

Test that WebVTT objects are being styled correctly based on user settings that should override default settings.
EVENT(canplaythrough)

EXPECTED (getComputedStyle(cue).color == 'rgb(255, 255, 255)') OK
EXPECTED (getComputedStyle(cue).backgroundColor == 'rgba(0, 0, 0, 0.8)') OK
EXPECTED (getComputedStyle(cue).fontFamily == 'sans-serif') OK
RUN(video.currentTime = 0.3)

EXPECTED (getComputedStyle(cue).color == 'rgb(128, 0, 128)') OK
EXPECTED (getComputedStyle(cue).fontSize == '14px') OK
EXPECTED (getComputedStyle(cue).backgroundColor == 'rgba(0, 0, 0, 0.8)') OK
EXPECTED (getComputedStyle(cue).fontFamily == 'sans-serif') OK
END OF TEST
